+++ project 2/API/Responses/TransactionResponse.swift
@@ -11,7 +11,7 @@ struct TransactionResponse: Codable {
     let id: String
     let senderId: String
     let receiverId: String
-    let amount: Double
+    let amount: Decimal
     let currency: String
     let createdOn: Int
     let reference: String
